requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0
langchain-aws>=0.1.0
langchain-core>=0.1.14
langsmith>=0.0.70
boto3>=1.34.0
pymongo>=4.6.1
python-dotenv>=1.0.0
pyjwt